Mediator sample refactoring and smell removal
